[ALPS07414878]  Disp: Fix coverity issue
ALPS07414878
ALPS07415327
ALPS07414934
ALPS07415326
Dereference null return value

ALPS07415241
ALPS07415303
Out-of-bounds access

ALPS07415035
leaked_storage

ALPS07415318
ALPS07415590
an address that could be at negative offset of an array

ALPS07415682
ALPS07415109
ALPS07415129
ALPS07415148
ALPS07415567
Dereference before null check

ALPS07415028
Division or modulo by zero

ALPS07433056
ALPS07415420
Unintentional integer overflow

ALPS07414907
Dereference null pointer

MTK-Commit-Id: f150c1f25bf01a2aa553211d1c5c115dcc094cc6

CR-Id: ALPS07414878
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_dp_intf.c
@@ -890,7 +890,7 @@ unsigned long long mtk_dpintf_get_frame_hrt_bw_base(
 	int vrefresh = drm_mode_vrefresh(&mtk_crtc->base.state->adjusted_mode);
 	u32 bpp = 3;
 
-	bw_base = vact * hact * vrefresh * 4 / 1000;
+	bw_base = (unsigned long long)vact * hact * vrefresh * 4 / 1000;
 	bw_base = bw_base * vtotal / vact;
 	bw_base = bw_base / 1000;
